% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badgen.R
\name{badgen}
\alias{badgen}
\title{Generate badges}
\usage{
badgen(
  label,
  status = "v1.2.3",
  color = "green",
  labelColor = "555",
  style = "classic",
  scale = 1,
  svg_icon = NULL
)
}
\arguments{
\item{label}{text for left part of the badge}

\item{status}{text for right part of the badge}

\item{color}{background for status part}

\item{labelColor}{background for label part}

\item{style}{one of 'classic' or 'flat'}

\item{scale}{resize badge, default size is 1}
}
\description{
Generate badges
}
\examples{
svg <- badgen("yolo", "success!", style = 'flat')
\dontrun{
writeLines(svg, 'test.svg')
browseURL('test.svg')

# Convert svg to bitmap
rsvg::rsvg_png('test.svg', 'test.png')
browseURL('test.png')}

# Add an svg icon to the badge
rlogo <- readLines('https://www.r-project.org/logo/Rlogo.svg')
svg2 <- badgen('mypkg', 'awesome', scale = 4, svg_icon = rlogo)

\dontrun{
writeLines(svg2, 'test2.svg')
browseURL('test2.svg')
rsvg::rsvg_png('test2.svg', 'test2.png')
browseURL('test2.png')}
}
